third_party/WebKit/LayoutTests/http/tests/serviceworker/postmessage.html

<!DOCTYPE html>
<title>Service Worker: postMessage</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
var test = async_test('postMessage to a ServiceWorker (and back via MessagePort)');
test.step(function() {

    navigator.serviceWorker.register('resources/postmessage-worker.js').then(
        test.step_func(function(worker) {
            var messageChannel = new MessageChannel();
            messageChannel.port1.onmessage = test.step_func(onMessage);

            worker.postMessage({port: messageChannel.port2}, [messageChannel.port2]);

            worker.postMessage({value: 1});
            worker.postMessage({value: 2});
            worker.postMessage({done: true});
        }),
        test.step_func(function(reason) {
            assert_unreached('Registration should succeed, but failed: ' + reason.name);
        }));

    var result = [];
    var expected = [
        'Acking value: 1',
        'Acking value: 2',
    ];

    function onMessage(e) {
        var message = e.data;
        if (message === 'quit') {
            assert_array_equals(result, expected, 'Worker should post back expected values.');
            test.done();
        } else {
            result.push(message);
        }
    };
});
</script>
